+++ Dupla 15/projeto-loterica/Loterica.Infra.Data/Features/Resultados/ResultadoRepository.cs
@@ -1,19 +1,17 @@
-﻿using Loterica.Dominio.Base;
+﻿using Loterica.Dominio.Exceptions;
 using Loterica.Dominio.Features.Resultados;
 using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Loterica.Infra.Data.Features.Resultados
 {
-    public class ResultadoRepository : IRepository<Resultado>
+    public class ResultadoRepository : IResultadoRepository
     {
         #region Querys
-        private const string _insert = @"INSERT INTO Resultado (NumerosSorteados) VALUES (@NumerosSorteados)";
-        private const string _update = @"UPDATE Resultado SET NumerosSorteados = @NumerosSorteados WHERE Id = @Id";
+        private const string _insert = @"INSERT INTO Resultado (NumerosSorteados, MediaQuadra, MediaQuina, MediaSena) VALUES (@NumerosSorteados, @MediaQuadra, @MediaQuina, @MediaSena);";
+        private const string _update = @"UPDATE Resultado SET NumerosSorteados = @NumerosSorteados, MediaQuadra = @MediaQuadra, MediaQuina = @MediaQuina, MediaSena = @MediaSena WHERE Id = @Id";
         private const string _getById = @"SELECT * FROM Resultado WHERE Id = @Id";
         private const string _getAll = @"SELECT * FROM Resultado";
         private const string _delete = @"DELETE FROM Resultado WHERE Id = @Id";
@@ -21,15 +19,22 @@ namespace Loterica.Infra.Data.Features.Resultados
 
         public Resultado Adicionar(Resultado entidade)
         {
+            entidade.Validar();
+
             long id = Db.Insert(_insert, Take(entidade));
 
-            entidade = ObterPorId(Convert.ToInt32(id));
+            entidade = ObterPorId(id);
 
             return entidade;
         }
 
         public Resultado Atualizar(Resultado entidade)
         {
+            if (entidade.Id == 0)
+                throw new IdentifierUndefinedException();
+
+            entidade.Validar();
+
             Db.Update(_update, Take(entidade));
 
             entidade = ObterPorId(entidade.Id);
@@ -39,11 +44,19 @@ namespace Loterica.Infra.Data.Features.Resultados
 
         public void Deletar(Resultado entidade)
         {
+            if (entidade.Id == 0)
+                throw new IdentifierUndefinedException();
+
+            entidade.Validar();
+
             Db.Delete(_delete, Take(entidade));
         }
 
-        public Resultado ObterPorId(int id)
+        public Resultado ObterPorId(long id)
         {
+            if (id == 0)
+                throw new IdentifierUndefinedException();
+            
             return Db.Get(_getById, Make, new object[] { "@Id", id });
         }
 
@@ -55,8 +68,11 @@ namespace Loterica.Infra.Data.Features.Resultados
         private static Func<IDataReader, Resultado> Make = reader =>
            new Resultado
            {
-               Id = Convert.ToInt32(reader["Id"]),
-               NumerosSorteados = (Convert.ToString(reader["NumerosSorteados"])).Split(',').Select(Int32.Parse).ToList()
+               Id = Convert.ToInt64(reader["Id"]),
+               NumerosSorteados = (Convert.ToString(reader["NumerosSorteados"])).Split(',').Select(Int32.Parse).ToList(),
+               MediaQuadra = Convert.ToDecimal(reader["MediaQuadra"]),
+               MediaQuina = Convert.ToDecimal(reader["MediaQuina"]),
+               MediaSena = Convert.ToDecimal(reader["MediaSena"])
            };
 
         private object[] Take(Resultado resultado)
@@ -64,7 +80,10 @@ namespace Loterica.Infra.Data.Features.Resultados
             return new object[]
             {
                 "@Id", resultado.Id,
-                "@NumerosSorteados", string.Join(",", resultado.NumerosSorteados.Select(n => n.ToString()).ToArray())
+                "@NumerosSorteados", string.Join(",", resultado.NumerosSorteados.Select(n => n.ToString()).ToArray()),
+                "@MediaQuadra", resultado.MediaQuadra,
+                "@MediaQuina", resultado.MediaQuina,
+                "@MediaSena", resultado.MediaSena
             };
         }
     }
